+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Restart fwcloud service after update it.
+- Init fwcloud-ui submodules.
+
 
 ## [1.0.4] - 2021-01-27
 ### Added
@@ -17,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Location of .env file.
 - Removed database selection and install MariaDB if no database engine installed. 
 - Improve database type detection.
+
 
 ## [1.0.3] - 2020-12-01
 ### Added